[Cornia/config] Remove unnecessary defines & move all callbacks to keymap.c
