--- keyboards/cornia/keymaps/fire/callback_system.c
@@ -1,31 +1,0 @@
-/* Copyright 2024 Vaarai
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 2 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-#include QMK_KEYBOARD_H
-
-#include "./keymap.h"
-
-bool shutdown_user(bool jump_to_bootloader) {
-    oled_clear();
-    oled_set_cursor(0, 2);
-    if (jump_to_bootloader) {
-        oled_write_P(PSTR("FLASH"), false);
-    } else {
-        oled_write_P(PSTR("RESET"), false);
-    }
-    oled_render_dirty(true);
-    return false;
-}
