+++ keyboards/cornia/keymaps/fire/keymap.h
@@ -18,6 +18,10 @@
 
 #include QMK_KEYBOARD_H
 
+/* Trackpad srolling speed adjustment */
+#define SCROLL_DIVISOR_H 8.0
+#define SCROLL_DIVISOR_V 8.0
+
 /* Trackpad srolling enablement flag */
 extern bool set_scrolling;
