+++ keyboards/cornia/keymaps/fire/rules.mk
@@ -13,20 +13,11 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-SRC += callback_oled.c
-SRC += callback_pointing_device.c
-SRC += callback_record.c
-SRC += callback_system.c
 SRC += tap_dances.c
 
 MOUSEKEY_ENABLE = yes
 CAPS_WORD_ENABLE = yes
 TAP_DANCE_ENABLE = yes
-
-# OLED I²C configuration
-OLED_ENABLE = yes
-OLED_DRIVER = ssd1306
-OLED_TRANSPORT = i2c
 
 # Cirque Trackpad I²C configuration
 POINTING_DEVICE_ENABLE = yes
