+++ keyboards/cornia/v1/keyboard.json
@@ -9,6 +9,7 @@
     "features": {
         "extrakey": true,
         "nkro": true,
+        "oled": true,
         "rgb_matrix": true
     },
     "split": {
